requirements.txt
# Core dependencies for Clash Royale Bot
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
pyautogui==0.9.54
pynput==1.7.6
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
pyyaml==6.0.1
loguru==0.7.0
tqdm==4.65.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2

# Android emulator specific
adb-shell==0.4.3
pure-python-adb==0.3.0.dev0
